cove: Rework of results processing in cove
This is a rework of the results processing in cove to allow for much
larger files. It uses the existing architecture of cove which relies on
the jobs being executed by the django view.

This is something to be reworked more fully when we have more available
time into a system using separate worker processes.