tests_functional: Remove invalid codelist warning for null
This is now failing due to a quirk in the way in which null is
interpreted when not in an embedded codelist. The lack of message here
is consistent with other codelists. This behaviour changed as a result
of using a local version of the grant standard and downloaded codelist
when "resolving" the standard.

https://github.com/ThreeSixtyGiving/standard/issues/379