tests: Add basic test for file submission and Fix CSRF issue
Adds new test to check that the basic file submission process works.

Due to the pending upgrade to lib-cove-web and the use of pytest instead
of django's test runner we also need to fix the use of ?source_url as
a parameter used in the input tests. This is done via a
new settings file for tests and switching ON `allow_direct_web_fetch`.

This isn't ideal as it isn't enabled in live however rewriting all the
tests which use this feature or other strategies such as
@override_settings requires refactoring which is out of scope.